requirements.txt
jmespath>=0.10.0
python-slugify>=6.1.1
jsonschema>=4.17.0
dnspython>=2.2.1
pyyaml>=6.0.1
Markdown>=3.4.4
jwt>=1.3.1
pygithub>=2.1.1
typer>=0.9.0 # CLI
rich>=13.7.0 # pretty print for typer
pyotp>=2.9.0
s3cmd>=2.4.0
azure-data-tables>=12.5.0
hvac>=2.2.0
